# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: REST API для управления банковскими картами
  contact:
    name: noname-jack
    email: aleksander.moskovchenko@mail.ru
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/transfer/my:
    post:
      tags:
        - Переводы
      summary: Перевод между своими картами (для пользователя)
      description: "Выполняет перевод средств между картами, принадлежащими аутент\
        ифицированному пользователю."
      operationId: transferBetweenMyCards
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: Перевод успешно выполнен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoTransferResponse"
  /api/cards:
    get:
      tags:
        - Карты
      summary: Получение списка всех карт (для администратора)
      description: Возвращает постраничный список всех карт с возможностью фильтрации.
        Доступно только администраторам.
      operationId: getAllCardsForAdmin
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
        - name: filter
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FilterCardAdminRequest"
      responses:
        "200":
          description: Список карт успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoPagedModelCardResponse"
    post:
      tags:
        - Карты
      summary: Создание новой карты (для администратора)
      description: Создаёт новую карту на основе переданных данных. Доступно только
        администраторам.
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "201":
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardDetailedResponse"
  /api/cards/my/{cartId}/block-request:
    post:
      tags:
        - Карты
      summary: Создание запроса на блокировку своей карты (для пользователя)
      description: "Создаёт запрос на блокировку карты, принадлежащей пользователю\
        . Доступно только аутентифицированным пользователям."
      operationId: createBlockRequest
      parameters:
        - name: cartId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "201":
          description: Запрос на блокировку успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
        "404":
          description: Карта не найдена или не принадлежит пользователю
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
...
  /api/cards/{id}/block:
    patch:
      tags:
        - Карты
      summary: Блокировка карты по ID (для администратора)
      description: "Блокирует карту, устанавливая статус BLOCKED. Доступно только\
        \ администраторам."
      operationId: blockCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: withResult
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Карта успешно заблокирована (с возвратом данных)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardDetailedResponse"
        "204":
          description: Карта успешно заблокирована (без возврата данных)
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
  /api/cards/{id}/activate:
    patch:
      tags:
        - Карты
      summary: Активация карты по ID (для администратора)
      description: "Активирует карту, устанавливая статус ACTIVE. Доступно только\
        \ администраторам."
      operationId: activateCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: withResult
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Карта успешно активирована (с возвратом данных)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardDetailedResponse"
        "204":
          description: Карта успешно активирована (без возврата данных)
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
  /api/cards/{cardId}:
    get:
      tags:
        - Карты
      summary: Получение детальной информации о карте по ID (для администратора)
      description: Получение детальной информации о карте по её идентификатору. Доступно
        только администраторам.
      operationId: getCardByIdForAdmin
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            exclusiveMinimum: 0
      responses:
        "200":
          description: Карта успешно найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardDetailedResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
  /api/cards/my:
    get:
      tags:
        - Карты
      summary: Получение списка своих карт (для пользователя)
      description: "Возвращает постраничный список карт, принадлежащих пользовател\
        ю, с возможностью фильтрации. Доступно только аутентифицированным пользова\
        телям."
      operationId: getMyCards
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
        - name: filter
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FilterCardUserRequest"
      responses:
        "200":
          description: Список карт успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoPagedModelCardResponse"
  /api/cards/my/{cardId}:
    get:
      tags:
        - Карты
      summary: Получение детальной информации о своей карте по ID
      description: Возвращает подробные данные о карте пользователя по её ID. Доступно
        только аутентифицированным пользователям.
      operationId: getMyCardById
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            exclusiveMinimum: 0
      responses:
        "200":
          description: Карта успешно найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardDetailedResponse"
        "404":
          description: Карта не найдена или не принадлежит пользователю
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
...
  /api/cards/{id}:
    delete:
      tags:
        - Карты
      summary: Удаление карты по ID (для администратора)
      description: Удаляет карту по её ID. Доступно только администраторам.
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Карта успешно удалена
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
components:
  schemas:
...
    TransferRequest:
      type: object
      description: Данные для перевода между картами
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          description: Сумма перевода
          example: 100.5
          maximum: 100000.00
          minimum: 1
      required:
        - amount
        - fromCardId
        - toCardId
    ApiResponseDtoTransferResponse:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          $ref: "#/components/schemas/TransferResponse"
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
...
    CardRequest:
      type: object
      description: Данные для создания новой карты
      properties:
        userId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2026-12-31
        balance:
          type: number
          description: Баланс карты
          example: 1000.5
      required:
        - balance
        - expirationDate
        - userId
    ApiResponseDtoCardDetailedResponse:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          $ref: "#/components/schemas/CardDetailedResponse"
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    FilterCardAdminRequest:
      type: object
      description: Данные для фильтрации карт администратором
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        ownerId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
        ownerUserName:
          type: string
          description: Имя пользователя владельца карты
          example: admin
        cardStatus:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balanceMin:
          type: number
          description: Минимальный баланс карты
          example: 0.0
        balanceMax:
          type: number
          description: Максимальный баланс карты
          example: 10000.0
    ApiResponseDtoPagedModelCardResponse:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          $ref: "#/components/schemas/PagedModelCardResponse"
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
    PagedModelCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
    FilterCardUserRequest:
      type: object
      description: Данные для фильтрации карт пользователя
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardStatus:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balanceMin:
          type: number
          description: Минимальный баланс карты
          example: 0.0
        balanceMax:
          type: number
          description: Максимальный баланс карты
          example: 10000.0
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT токен авторизации
      scheme: bearer
      bearerFormat: JWT